# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AgentGenerateBody {
  /** Plain-English description of the API to generate */
  prompt: string;
}

export interface GenerationRecord {
  id: number;
  prompt: string;
  filename: string;
  tokenCountPrompt?: number | null;
  tokenCountCompletion?: number | null;
  costUsd?: number | null;
  ttftMs?: number | null;
  modelUsed?: string | null;
  cacheHit?: boolean | null;
  createdAt: string;
}

export interface GatewayStats {
  /** Total number of generation requests */
  totalRequests: number;
  /** Number of requests served from semantic cache */
  cacheHits: number;
  /** Cache hit rate as a fraction (0-1) */
  cacheHitRate: number;
  /** Total tokens consumed (prompt + completion) */
  totalTokens: number;
  /** Total estimated cost in USD */
  totalCostUsd: number;
  /** Average time to first token in milliseconds (non-cached only) */
  avgTtftMs: number | null;
  /** Tokens consumed in the current 60-second window */
  tpmWindowTotal: number;
  /** Configured tokens-per-minute limit */
  tpmLimit: number;
}

export interface AuthUser {
  id: string;
  email?: string | null;
  firstName?: string | null;
  lastName?: string | null;
  profileImageUrl?: string | null;
}

export interface GetCurrentAuthUserResponse {
  user: AuthUser | null;
}

export interface ExchangeMobileAuthorizationCodeBody {
  code: string;
  code_verifier: string;
  redirect_uri: string;
  state: string;
  nonce?: string | null;
}

export interface ExchangeMobileAuthorizationCodeResponse {
  token: string;
}

export interface LogoutMobileSessionResponse {
  success: boolean;
}

export interface ApiKeyRecord {
  id: number;
  name: string;
  keyPrefix: string;
  monthlyLimit: number;
  revokedAt?: string | null;
  lastUsedAt?: string | null;
  createdAt: string;
}

export interface CreateApiKeyBody {
  /** Human-readable name for the key */
  name: string;
  /** Max requests per month (default 100) */
  monthlyLimit?: number;
}

export interface CreateApiKeyResponse {
  id: number;
  name: string;
  keyPrefix: string;
  /** Full API key — shown only once, store it now */
  key: string;
  monthlyLimit: number;
  createdAt: string;
}

export interface KeyUsageStats {
  monthYear: string;
  tokenCount: number;
  costUsd: number;
  requestCount: number;
  monthlyLimit: number;
}

export interface BillingPlan {
  id: string;
  name: string;
  description: string;
  priceId?: string | null;
  unitAmount?: number | null;
  currency: string;
  interval?: string | null;
  generationsPerMonth?: number | null;
}

export interface SubscriptionStatus {
  status: string | null;
  planName?: string | null;
  currentPeriodEnd?: string | null;
  cancelAtPeriodEnd?: boolean | null;
}

export interface CreateCheckoutBody {
  priceId: string;
}

export interface CheckoutResponse {
  url: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface ErrorResponse {
  error: string;
  code?: string;
}