# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: agent
    description: AI code generation operations
  - name: gateway
    description: Layer 8 Intelligence Gateway operations
  - name: auth
    description: Authentication operations
  - name: keys
    description: API key management
  - name: billing
    description: Billing and subscription management
  - name: public
    description: Public API gateway (API-key authenticated)
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /agent/generate:
    post:
      operationId: agentGenerate
      tags: [agent]
      summary: Generate async API code
      description: |
        Accepts a plain-English prompt and streams the AI-generated
        TypeScript async Express route handler as Server-Sent Events.
        On completion, the code is saved as a gzip-compressed file
        in the Agent/ directory. PII is redacted before sending to
        the model. Semantic cache is checked first.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentGenerateBody"
      responses:
        "200":
          description: SSE stream of generated code tokens
          content:
            text/event-stream:
              schema:
                type: string
        "400":
          description: Bad request or prompt injection detected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Token-per-minute rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /agent/history:
    get:
      operationId: agentHistory
      tags: [agent]
      summary: List past generations
      description: Returns all previously generated files, newest first.
      responses:
        "200":
          description: List of generation records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GenerationRecord"

  /gateway/stats:
    get:
      operationId: gatewayStats
      tags: [gateway]
      summary: Gateway aggregate statistics
      description: |
        Returns aggregate statistics for the Layer 8 gateway including
        cache hit rate, total token usage, estimated cost, and TTFT.
      responses:
        "200":
          description: Gateway statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewayStats"

  /auth/user:
    get:
      operationId: getCurrentAuthUser
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current auth state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetCurrentAuthUserResponse"

  /mobile-auth/token-exchange:
    post:
      operationId: exchangeMobileAuthorizationCode
      tags: [auth]
      summary: Exchange mobile authorization code for session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExchangeMobileAuthorizationCodeBody"
      responses:
        "200":
          description: Session token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExchangeMobileAuthorizationCodeResponse"

  /mobile-auth/logout:
    post:
      operationId: logoutMobileSession
      tags: [auth]
      summary: Logout mobile session
      responses:
        "200":
          description: Logout result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogoutMobileSessionResponse"

  /keys:
    get:
      operationId: listApiKeys
      tags: [keys]
      summary: List user's API keys
      responses:
        "200":
          description: List of API keys
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApiKeyRecord"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createApiKey
      tags: [keys]
      summary: Create a new API key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApiKeyBody"
      responses:
        "201":
          description: Newly created API key (plaintext shown once)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateApiKeyResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /keys/{id}:
    delete:
      operationId: revokeApiKey
      tags: [keys]
      summary: Revoke an API key
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Key revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Key not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /keys/{id}/usage:
    get:
      operationId: getApiKeyUsage
      tags: [keys]
      summary: Get monthly usage for an API key
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Monthly usage stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KeyUsageStats"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /billing/plans:
    get:
      operationId: getBillingPlans
      tags: [billing]
      summary: List available subscription plans
      responses:
        "200":
          description: Available plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BillingPlan"

  /billing/subscription:
    get:
      operationId: getBillingSubscription
      tags: [billing]
      summary: Get current user's subscription status
      responses:
        "200":
          description: Subscription info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionStatus"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /billing/checkout:
    post:
      operationId: createCheckoutSession
      tags: [billing]
      summary: Create a Stripe checkout session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCheckoutBody"
      responses:
        "200":
          description: Checkout session URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /billing/portal:
    post:
      operationId: createBillingPortalSession
      tags: [billing]
      summary: Create a Stripe billing portal session
      responses:
        "200":
          description: Portal session URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    AgentGenerateBody:
      type: object
      properties:
        prompt:
          type: string
          description: Plain-English description of the API to generate
      required:
        - prompt

    GenerationRecord:
      type: object
      properties:
        id:
          type: integer
        prompt:
          type: string
        filename:
          type: string
        tokenCountPrompt:
          type: integer
          nullable: true
        tokenCountCompletion:
          type: integer
          nullable: true
        costUsd:
          type: number
          nullable: true
        ttftMs:
          type: integer
          nullable: true
        modelUsed:
          type: string
          nullable: true
        cacheHit:
          type: boolean
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - prompt
        - filename
        - createdAt

    GatewayStats:
      type: object
      properties:
        totalRequests:
          type: integer
          description: Total number of generation requests
        cacheHits:
          type: integer
          description: Number of requests served from semantic cache
        cacheHitRate:
          type: number
          description: Cache hit rate as a fraction (0-1)
        totalTokens:
          type: integer
          description: Total tokens consumed (prompt + completion)
        totalCostUsd:
          type: number
          description: Total estimated cost in USD
        avgTtftMs:
          type: number
          nullable: true
          description: Average time to first token in milliseconds (non-cached only)
        tpmWindowTotal:
          type: integer
          description: Tokens consumed in the current 60-second window
        tpmLimit:
          type: integer
          description: Configured tokens-per-minute limit
      required:
        - totalRequests
        - cacheHits
        - cacheHitRate
        - totalTokens
        - totalCostUsd
        - avgTtftMs
        - tpmWindowTotal
        - tpmLimit

    AuthUser:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          nullable: true
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        profileImageUrl:
          type: string
          nullable: true
      required:
        - id

    GetCurrentAuthUserResponse:
      type: object
      properties:
        user:
          nullable: true
          oneOf:
            - $ref: "#/components/schemas/AuthUser"
            - type: "null"
      required:
        - user

    ExchangeMobileAuthorizationCodeBody:
      type: object
      properties:
        code:
          type: string
        code_verifier:
          type: string
        redirect_uri:
          type: string
        state:
          type: string
        nonce:
          type: string
          nullable: true
      required:
        - code
        - code_verifier
        - redirect_uri
        - state

    ExchangeMobileAuthorizationCodeResponse:
      type: object
      properties:
        token:
          type: string
      required:
        - token

    LogoutMobileSessionResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success

    ApiKeyRecord:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        keyPrefix:
          type: string
        monthlyLimit:
          type: integer
        revokedAt:
          type: string
          format: date-time
          nullable: true
        lastUsedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - keyPrefix
        - monthlyLimit
        - createdAt

    CreateApiKeyBody:
      type: object
      properties:
        name:
          type: string
          description: Human-readable name for the key
        monthlyLimit:
          type: integer
          description: Max requests per month (default 100)
      required:
        - name

    CreateApiKeyResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        keyPrefix:
          type: string
        key:
          type: string
          description: Full API key — shown only once, store it now
        monthlyLimit:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - keyPrefix
        - key
        - monthlyLimit
        - createdAt

    KeyUsageStats:
      type: object
      properties:
        monthYear:
          type: string
        tokenCount:
          type: integer
        costUsd:
          type: number
        requestCount:
          type: integer
        monthlyLimit:
          type: integer
      required:
        - monthYear
        - tokenCount
        - costUsd
        - requestCount
        - monthlyLimit

    BillingPlan:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        priceId:
          type: string
          nullable: true
        unitAmount:
          type: integer
          nullable: true
        currency:
          type: string
        interval:
          type: string
          nullable: true
        generationsPerMonth:
          type: integer
          nullable: true
      required:
        - id
        - name
        - description
        - currency

    SubscriptionStatus:
      type: object
      properties:
        status:
          type: string
          nullable: true
        planName:
          type: string
          nullable: true
        currentPeriodEnd:
          type: string
          format: date-time
          nullable: true
        cancelAtPeriodEnd:
          type: boolean
          nullable: true
      required:
        - status

    CreateCheckoutBody:
      type: object
      properties:
        priceId:
          type: string
      required:
        - priceId

    CheckoutResponse:
      type: object
      properties:
        url:
          type: string
      required:
        - url

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
      required:
        - error