# lib/api-zod/src/generated/types/apiKeyRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ApiKeyRecord {
  id: number;
  name: string;
  keyPrefix: string;
  monthlyLimit: number;
  revokedAt?: Date | null;
  lastUsedAt?: Date | null;
  createdAt: Date;
}